+++ src/LuaState.cpp
@@ -124,6 +124,9 @@ void LuaState::open_libraries(BitField<Library> libraries) {
 		if (libraries.has_flag(GODOT_ENUMS)) {
 			lua_state.require(module_names::enums, &luaopen_godot_enums, false);
 		}
+		if (libraries.has_flag(GODOT_LOCAL_PATHS)) {
+			lua_state.require(module_names::local_paths, &luaopen_godot_local_paths, false);
+		}
 	}
 }
 
@@ -131,20 +134,28 @@ Ref<LuaTable> LuaState::create_table(const Dictionary& initial_values) {
 	return memnew(LuaTable(to_table(lua_state, initial_values)));
 }
 
-Variant LuaState::load_string(const String& chunk, const String& chunkname, LuaTable *env) {
-	return ::luagdextension::load_string(lua_state, chunk, chunkname, env);
+Variant LuaState::load_buffer(const PackedByteArray& chunk, const String& chunkname, LoadMode mode, LuaTable *env) {
+	return ::luagdextension::load_buffer(lua_state, chunk, chunkname, (sol::load_mode) mode, env);
 }
 
-Variant LuaState::load_file(const String& filename, int buffer_size, LuaTable *env) {
-	return ::luagdextension::load_file(lua_state, filename, buffer_size, env);
+Variant LuaState::load_string(const String& chunk, const String& chunkname, LuaTable *env) {
+	return ::luagdextension::load_buffer(lua_state, chunk.to_utf8_buffer(), chunkname, sol::load_mode::text, env);
+}
+
+Variant LuaState::load_file(const String& filename, LoadMode mode, LuaTable *env) {
+	return ::luagdextension::load_file(lua_state, filename, (sol::load_mode) mode, env);
+}
+
+Variant LuaState::do_buffer(const PackedByteArray& chunk, const String& chunkname, LoadMode mode, LuaTable *env) {
+	return ::luagdextension::do_buffer(lua_state, chunk, chunkname, (sol::load_mode) mode, env);
 }
 
 Variant LuaState::do_string(const String& chunk, const String& chunkname, LuaTable *env) {
-	return ::luagdextension::do_string(lua_state, chunk, chunkname, env);
+	return ::luagdextension::do_buffer(lua_state, chunk.to_utf8_buffer(), chunkname, sol::load_mode::text, env);
 }
 
-Variant LuaState::do_file(const String& filename, int buffer_size, LuaTable *env) {
-	return ::luagdextension::do_file(lua_state, filename, buffer_size, env);
+Variant LuaState::do_file(const String& filename, LoadMode mode, LuaTable *env) {
+	return ::luagdextension::do_file(lua_state, filename, (sol::load_mode) mode, env);
 }
 
 LuaTable *LuaState::get_globals() const {
@@ -187,17 +198,25 @@ void LuaState::_bind_methods() {
 	BIND_BITFIELD_FLAG(GODOT_SINGLETONS);
 	BIND_BITFIELD_FLAG(GODOT_CLASSES);
 	BIND_BITFIELD_FLAG(GODOT_ENUMS);
+	BIND_BITFIELD_FLAG(GODOT_LOCAL_PATHS);
 	BIND_BITFIELD_FLAG(GODOT_ALL_LIBS);
 	
 	BIND_BITFIELD_FLAG(ALL_LIBS);
 
+	// LoadMode enum
+	BIND_ENUM_CONSTANT(LOAD_MODE_ANY);
+	BIND_ENUM_CONSTANT(LOAD_MODE_TEXT);
+	BIND_ENUM_CONSTANT(LOAD_MODE_BINARY);
+
 	// Methods
 	ClassDB::bind_method(D_METHOD("open_libraries", "libraries"), &LuaState::open_libraries, DEFVAL(BitField<Library>(ALL_LIBS)));
 	ClassDB::bind_method(D_METHOD("create_table", "initial_values"), &LuaState::create_table, DEFVAL(Dictionary()));
+	ClassDB::bind_method(D_METHOD("load_buffer", "chunk", "chunkname", "mode", "env"), &LuaState::load_buffer, DEFVAL(""), DEFVAL(LOAD_MODE_ANY), DEFVAL(nullptr));
 	ClassDB::bind_method(D_METHOD("load_string", "chunk", "chunkname", "env"), &LuaState::load_string, DEFVAL(""), DEFVAL(nullptr));
-	ClassDB::bind_method(D_METHOD("load_file", "filename", "buffer_size", "env"), &LuaState::load_file, DEFVAL(1024), DEFVAL(nullptr));
+	ClassDB::bind_method(D_METHOD("load_file", "filename", "mode", "env"), &LuaState::load_file, DEFVAL(LOAD_MODE_ANY), DEFVAL(nullptr));
+	ClassDB::bind_method(D_METHOD("do_buffer", "chunk", "chunkname", "mode", "env"), &LuaState::do_buffer, DEFVAL(""), DEFVAL(LOAD_MODE_ANY), DEFVAL(nullptr));
 	ClassDB::bind_method(D_METHOD("do_string", "chunk", "chunkname", "env"), &LuaState::do_string, DEFVAL(""), DEFVAL(nullptr));
-	ClassDB::bind_method(D_METHOD("do_file", "filename", "buffer_size", "env"), &LuaState::do_file, DEFVAL(1024), DEFVAL(nullptr));
+	ClassDB::bind_method(D_METHOD("do_file", "filename", "mode", "env"), &LuaState::do_file, DEFVAL(LOAD_MODE_ANY), DEFVAL(nullptr));
 	ClassDB::bind_method(D_METHOD("get_globals"), &LuaState::get_globals);
 	ClassDB::bind_method(D_METHOD("get_registry"), &LuaState::get_registry);