+++ test/lua_tests/dofile_res.lua
@@ -1,0 +1,3 @@
+dofile("res://lua_tests/return_args.lua")
+
+assert(not pcall(function() dofile("res://non-existent-file") end))