+++ test/lua_tests/package_searcher.lua
@@ -1,0 +1,2 @@
+assert(package.searchpath("lua_tests.package_searcher", "res://?.lua"))
+assert(not package.searchpath("module.that.doesnt.exist", "res://?.lua"))